+++ tests/integrations/checks_test.py
@@ -19,7 +19,7 @@ class TestGetCustomChecks:
         config = configuration.Config(
             checks=configuration.Checks(
                 summary=(
-                    configuration.SummaryCheck(
+                    configuration.Check(
                         pattern=r'\d+',
                         message='Commit summary contains numbers',
                     ),
@@ -33,6 +33,27 @@ class TestGetCustomChecks:
             checks.SummaryRegexpCheck(
                 pattern=r'\d+',
                 message='Commit summary contains numbers',
+            ),
+        ]
+
+    def test_author_email_regexp_check(self) -> None:
+        config = configuration.Config(
+            checks=configuration.Checks(
+                author_email=(
+                    configuration.Check(
+                        pattern=r'@example\.com$',
+                        message='Fake email address provided.',
+                    ),
+                ),
+            ),
+        )
+
+        custom_checks = checks.get_custom_checks(config)
+
+        assert custom_checks == [
+            checks.AuthorEmailRegexpCheck(
+                pattern=r'@example\.com$',
+                message='Fake email address provided.',
             ),
         ]
 
@@ -54,6 +75,34 @@ class TestSummaryRegexpCheck:
         regexp_check = checks.SummaryRegexpCheck(
             pattern=r'\d+',
             message='Summary contains numbers.',
+        )
+
+        errors = regexp_check(commit)
+
+        assert list(errors) == []
+
+
+class TestAuthorEmailRegexpCheck:
+    def test_matches(self) -> None:
+        commit = mock.Mock(
+            spec_set=application.Commit, author_email='april.may@example.com'
+        )
+        regexp_check = checks.AuthorEmailRegexpCheck(
+            pattern=r'@example\.com$',
+            message='Fake email address provided.',
+        )
+
+        errors = regexp_check(commit)
+
+        assert list(errors) == ['Fake email address provided.']
+
+    def test_no_matches(self) -> None:
+        commit = mock.Mock(
+            spec_set=application.Commit, author_email='april.may@example.net'
+        )
+        regexp_check = checks.AuthorEmailRegexpCheck(
+            pattern=r'@example\.com$',
+            message='Fake email address provided.',
         )
 
         errors = regexp_check(commit)