+++ validate_commits/integrations/configuration.py
@@ -12,14 +12,15 @@ if TYPE_CHECKING:
 
 
 @attrs.frozen
-class SummaryCheck:
+class Check:
     pattern: str
     message: str
 
 
 @attrs.frozen
 class Checks:
-    summary: tuple[SummaryCheck, ...] = ()
+    summary: tuple[Check, ...] = ()
+    author_email: tuple[Check, ...] = ()
 
 
 @attrs.frozen